Added logic that will refresh root feed only when changes are detected from account metadata.
BUG=chromium-os:28820
TEST=manual - make sure directory refreshes are not happening every five minutes even if there were no changes on the server side.

Review URL: https://chromiumcodereview.appspot.com/9969108

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@130694 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/chromeos/gdata/gdata_parser.h
@@ -550,11 +550,11 @@ class AccountMetadataFeed {
   // class for the details.
   static AccountMetadataFeed* CreateFrom(base::Value* value);
 
-  int quota_bytes_total() const {
+  int64 quota_bytes_total() const {
     return quota_bytes_total_;
   }
 
-  int quota_bytes_used() const {
+  int64 quota_bytes_used() const {
     return quota_bytes_used_;
   }
 
@@ -574,8 +574,8 @@ class AccountMetadataFeed {
   // Return false if parsing fails.
   bool Parse(base::Value* value);
 
-  int quota_bytes_total_;
-  int quota_bytes_used_;
+  int64 quota_bytes_total_;
+  int64 quota_bytes_used_;
   int largest_changestamp_;
 
   static const char kQuotaBytesTotalField[];
